refactor(Collapse): move types and atom factories to separate files
